Added support for creating HDF5 files in Blaze constructors

+++ blaze/tests/common.py
@@ -13,20 +13,28 @@ import glob
 class MayBePersistentTest():
 
     disk = False
+    dir_ = False
 
     def setUp(self):
         if self.disk:
-            prefix = 'barray-' + self.__class__.__name__
-            self.rootdir = tempfile.mkdtemp(prefix=prefix)
-            os.rmdir(self.rootdir)  # tests needs this cleared
+            if self.dir_:
+                prefix = 'barray-' + self.__class__.__name__
+                self.rootdir = tempfile.mkdtemp(prefix=prefix)
+                os.rmdir(self.rootdir)  # tests needs this cleared
+            else:
+                handle, self.file = tempfile.mkstemp()
+                os.close(handle)  # close the non needed file handle
         else:
             self.rootdir = None
 
     def tearDown(self):
         if self.disk:
-            # Remove every directory starting with rootdir
-            for dir_ in glob.glob(self.rootdir+'*'):
-                shutil.rmtree(dir_)
+            if self.dir_:
+                # Remove every directory starting with rootdir
+                for dir_ in glob.glob(self.rootdir+'*'):
+                    shutil.rmtree(dir_)
+            else:
+                os.unlink(self.file)
 
 
 class BTestCase(unittest.TestCase):
